Bug fixed: wrong folding for block scheduling
When block schedules are scheduled on the same block period, the folding
function which translates from diary to calfw, can not distinguish them,
because the function make an alist by the schedule period as a key.
This patch corrects the folding function and the schedule items are
displayed as users expect. (thx abetatsuya)