Fix recurring events in single-org-file sources
Prior to this change, recurring events worked
in org-create-source, but NOT in org-create-file-source.
This is not the cleanest lisp code, because I am
cookbooking my way through making something work
and not properly understanding what I am doing,
BUT org-create-file-source NOW handles recurring
events correctly.

Ideally, instead of making separate sets of functions for
cfw:org-file-schedule-period-to-calendar and
cfw:org-file-collect-schedules-period
I should have made the "file" argument ":optional"
(in both original functions) and handled both
org-create-file-source and org-create-source
through the same chain of
cfw:org-schedule-period-to-calendar and
cfw:org-file-collect-schedules-period
... My lisp skills were not up to completing
that approach CLEANLY, so I have done things this way.