Feat: Add a variable to control whether face-today is applied
This variable gives the user control on whether `cfw:face-today` is
applied to the whole day view or not. The default value `t` ensures
that existing behaviour is not affected.

I use `emacs-calfw` in Day view mode, and I do not want the
`cfw:face-today` background in Day View. I wouldn't mind it in Week
View. Hiding the application of the face behind a variable gives me
the control I prefer.